Fixed Bug - Ratpack request delegates not thread safe
'params' and 'urlparams' are not thread safe.
If a second request comes into a request handler while the previous one is
still processing, the value of 'params' and 'urlparams' are replaced in the
namespace of the original request.

The fix is to clone the request handler closure before setting the delegate,
thus creating a localized namespace.

This will impart a small performance hit